Add files via upload
Working on this Bot|Script for OSRS, I discovered away to force the programming to be false, allowing me to go through each while, by force on the way out of the file.

Every state is at first @ an OFF || FALSE state, once it goes into all the while loops to the Lumbridge location, it will choose a random task to do, after it finish and gets forced to the next location it turns OFF this spot and breaks out of all the while loops, after it does this. The bot structure will refire and come back into the while loops to Draynor, which at this point is @ a TRUE || ON state, while the rest is OFF before and after it. and so on forth.

Once a while loop becomes true the programming aka bot || script will then fire the block of code right under the while loop, if not this is needed in programming!!!!

I have not tested this at the current moment, I'm currently thinking I will need to put them inside the closing } of the while loop in order for this to happen, yet I'm sure it will do the same out side.

Still building it.

Also becomes a prefect firing sequence of a multi stage rocket, I will be working on a 3d blended file of this said rocket.

conceptually allowing the rocket to have X amount of shells one over the top of the other, allowing the rocket to dump it's guts as it climbs. YES multi stages each stage releasing from it's center point, after the fuel has emptied.

modify this said bot || script with sleep timer in between every failed while statement can be the example of one shell burning out and releasing from the reset of the rocket, The core being the main, the rocket will be beefy, allowing it to sustain  flight, and causing it to get lighter as it climbs, leaving behind reusable shells. Contact I'll be waiting!

The amount of structure required would be our current rockets x X shells one slightly bigger then the other, with the tip being able to be at the ozone before the last shell "Doesn't more fuel in space".

 /\
/  \
| /\|
/  \
| /\| core max, while each shell then becomes lesser & lesser. My mind believes 8 secs flat after flight X miles above ground maybe 5000.
/  \
| /\| yes each shell holds it's own fuel cells
/  \
 ||| represents fuel
not a full representation, yet a small one inside || outside the other.


Thanks for the mindset goes to Cheating on games sence I was C&C Unknown Cheats / DB / RuneMate / Jagex, OSRS, Runescape NEW